refactor: Text 관련 함수 공용화
